move LqasAfroViewset outside of common.py
