+++ plugins/polio/api/lqas_im/lqasim_global_map.py
@@ -5,7 +5,8 @@ from rest_framework.response import Response
 from iaso.models import OrgUnit
 from iaso.models.data_store import JsonDataStore
 from iaso.utils import geojson_queryset
-from plugins.polio.api.common import LqasAfroViewset, calculate_country_status, LQASStatus, RoundSelection
+from plugins.polio.api.common import LQASStatus, RoundSelection, calculate_country_status
+from plugins.polio.api.lqas_im.base_viewset import LqasAfroViewset
 from plugins.polio.api.lqas_im.lqasim_zoom_in_map import get_latest_active_campaign_and_rounds
